refactor: apply batch-01 fixes (issues ISSUE-0001 to ISSUE-0020)
- Linter versions: rustc 1.87.0-nightly, clippy 0.1.87
- Rule references: clippy::redundant_closure, clippy::unnecessary_cast, clippy::absurd_extreme_comparisons
- Test coverage: 582 tests passed

+++ src/tests/test_bag_system.rs
@@ -331,7 +331,7 @@ fn test_bag_distribution_statistics() {
     let expected: f32 = 100.0; // 700 / 7
     let variance: f32 = counts
         .iter()
-        .map(|&c| ((c as f32 - expected).powi(2)) as f32)
+        .map(|&c| (c as f32 - expected).powi(2))
         .sum::<f32>()
         / 7.0;
     assert!(
